russian-handwriting-checker-new/python-service/requirements.txt
# Основные зависимости
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0
pydantic[email]>=2.0
python-dotenv>=1.0
openai>=1.57.0

# MarkItDown — главный конвертер
markitdown[all]
markitdown-ocr

# OCR
paddleocr==2.8.1
opencv-python==4.10.0.84
pytesseract
Pillow
numpy==1.26.4

# Дополнительно
python-multipart>=0.0.9

# Auth
python-jose[cryptography]>=3.3.0

# Supabase
supabase>=2.0.0

# DB — PostgreSQL (Supabase) / SQLite (локально)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Migrations
alembic>=1.13.0
